engine: it is now possible to end intermission with dedicated server; also, don't run VM thinkers if there are no active clients

+++ source/server.h
@@ -25,7 +25,7 @@
 //**************************************************************************
 struct server_t {
   int intermission;
-  int intertime;
+  float intertime;
 };
 
 struct server_static_t {
